refactor(uibackend): generate per module
