+++ Makefile
@@ -295,7 +295,9 @@ gen-apiserver-api: ## Generating Go library for API specification
 .PHONY: gen-uibackend-api
 gen-uibackend-api: ## Generating Go library for UI Backend API specification
 	$(info Generating API for UI backend code ...)
-	@(cd uibackend && go generate)
+	@(cd uibackend/client && go generate)
+	@(cd uibackend/server/pkg && go generate)
+	@(cd uibackend/types && go generate)
 
 .PHONY: gen-bicep
 gen-bicep: bin/bicep ## Generating Azure Bicep template(s)
