+++ uibackend/client/client.cfg.yaml
@@ -1,5 +1,5 @@
 package: client
-output: client/client.gen.go
+output: client.gen.go
 additional-imports:
   - package: github.com/openclarity/vmclarity/uibackend/types
     alias: .
