+++ uibackend/types/types.cfg.yaml
@@ -1,6 +1,6 @@
 package: types
 generate:
   models: true
-output: types/types.gen.go
+output: types.gen.go
 output-options:
   skip-prune: true
